feat: update sidebar navigation for orders management
refactor: streamline OrdersList component by integrating useCustomerOrders hook

feat: add order detail page for admin with actions to confirm, ship, and cancel orders

feat: implement admin order management page with filtering and pagination

feat: create useCustomerOrders hook for fetching customer orders with query support
